refactor(tests): split TransactionRepositoryContractTests (979→≤261) (#272, #266, #270)
Split MoolahTests/Domain/TransactionRepositoryContractTests.swift along
the // MARK: seams into five @Suite-per-file files plus a fixture
support file. Every new file is strictly under the 400-line
file_length threshold and every struct body is under 250 lines.

Also clears baseline entries that would otherwise re-key:
  - 77 × force_unwrapping → try #require(...) (tests made throws).
  - 2 × force_try → throwing factory helpers.
  - 1 × function_body_length (58 lines) → UpdateRoundTripFixture helper
    keeps the inline assertions (the point of the test) intact.
  - type_body_length + file_length eliminated by the split.

Split the suggested Persistence file into basic (Persistence) and
multi-instrument (MultiInst) suites to keep each struct body under 250.
Shortened `TransactionRepositoryAuxFilter` / `MultiInst` struct names to
stay under SwiftLint's 40-char type_name ceiling.

Stacked on #358 until it lands; retarget to main then.
