--- MoolahTests/Domain/TransactionRepositoryContractTests.swift
@@ -1,979 +1,0 @@
-import Foundation
-import SwiftData
-import Testing
-
-@testable import Moolah
-
-@Suite("TransactionRepository Contract")
-struct TransactionRepositoryContractTests {
-  @Test("filters by date range")
-  func testFiltersByDateRange() async throws {
-    let repository = makeCloudKitTransactionRepository(initialTransactions: makeTestTransactions())
-    let calendar = Calendar.current
-    let middleDate = calendar.date(from: DateComponents(year: 2024, month: 6, day: 15))!
-    let startDate = calendar.date(byAdding: .day, value: -30, to: middleDate)!
-    let endDate = calendar.date(byAdding: .day, value: 30, to: middleDate)!
-    let dateRange = startDate...endDate
-
-    let page = try await repository.fetch(
-      filter: TransactionFilter(dateRange: dateRange),
-      page: 0,
-      pageSize: 50
-    )
-
-    // Should only include transactions within the date range
-    for transaction in page.transactions {
-      #expect(dateRange.contains(transaction.date))
-    }
-  }
-
-  @Test("filters by category IDs")
-  func testFiltersByCategoryIds() async throws {
-    let repository = makeCloudKitTransactionRepository(initialTransactions: makeTestTransactions())
-    let transactions = makeTestTransactions()
-    let groceryCategory = transactions[0].legs.first(where: { $0.categoryId != nil })!.categoryId!
-    let categoryIds: Set<UUID> = [groceryCategory]
-
-    let page = try await repository.fetch(
-      filter: TransactionFilter(categoryIds: categoryIds),
-      page: 0,
-      pageSize: 50
-    )
-
-    // Should only include transactions with the specified category
-    for transaction in page.transactions {
-      #expect(transaction.legs.contains(where: { categoryIds.contains($0.categoryId ?? UUID()) }))
-    }
-  }
-
-  @Test("filters by payee (case-insensitive contains)")
-  func testFiltersByPayee() async throws {
-    let repository = makeCloudKitTransactionRepository(initialTransactions: makeTestTransactions())
-    let page = try await repository.fetch(
-      filter: TransactionFilter(payee: "wool"),
-      page: 0,
-      pageSize: 50
-    )
-
-    // Should include "Woolworths" (case-insensitive contains match)
-    #expect(!page.transactions.isEmpty)
-    for transaction in page.transactions {
-      let payee = transaction.payee?.lowercased() ?? ""
-      #expect(payee.contains("wool"))
-    }
-  }
-
-  @Test("combines multiple filters")
-  func testCombinesMultipleFilters() async throws {
-    let repository = makeCloudKitTransactionRepository(initialTransactions: makeTestTransactions())
-    let transactions = makeTestTransactions()
-    let groceryCategory = transactions[0].legs.first(where: { $0.categoryId != nil })!.categoryId!
-    let calendar = Calendar.current
-    let middleDate = calendar.date(from: DateComponents(year: 2024, month: 6, day: 15))!
-    let startDate = calendar.date(byAdding: .day, value: -30, to: middleDate)!
-    let endDate = calendar.date(byAdding: .day, value: 30, to: middleDate)!
-
-    let page = try await repository.fetch(
-      filter: TransactionFilter(
-        dateRange: startDate...endDate,
-        categoryIds: [groceryCategory],
-        payee: "wool"
-      ),
-      page: 0,
-      pageSize: 50
-    )
-
-    // Should satisfy all filter criteria
-    for transaction in page.transactions {
-      #expect(transaction.legs.contains(where: { $0.categoryId == groceryCategory }))
-      #expect((startDate...endDate).contains(transaction.date))
-      let payee = transaction.payee?.lowercased() ?? ""
-      #expect(payee.contains("wool"))
-    }
-  }
-
-  @Test("returns empty when no matches")
-  func testReturnsEmptyWhenNoMatches() async throws {
-    let repository = makeCloudKitTransactionRepository(initialTransactions: makeTestTransactions())
-    let page = try await repository.fetch(
-      filter: TransactionFilter(payee: "NonexistentPayee"),
-      page: 0,
-      pageSize: 50
-    )
-
-    #expect(page.transactions.isEmpty)
-  }
-
-  @Test("update preserves all transaction fields")
-  func testUpdatePreservesAllFields() async throws {
-    let repository = makeCloudKitTransactionRepository()
-    let calendar = Calendar.current
-    let fromAccountId = UUID()
-    let toAccountId = UUID()
-    let categoryId = UUID()
-    let earmarkId = UUID()
-    let date = calendar.date(from: DateComponents(year: 2024, month: 3, day: 12))!
-
-    let original = Transaction(
-      date: date,
-      payee: "Original Payee",
-      notes: "Some notes",
-      recurPeriod: .month,
-      recurEvery: 2,
-      legs: [
-        TransactionLeg(
-          accountId: fromAccountId, instrument: .defaultTestInstrument,
-          quantity: Decimal(string: "-750.00")!, type: .transfer,
-          categoryId: categoryId, earmarkId: earmarkId),
-        TransactionLeg(
-          accountId: toAccountId, instrument: .defaultTestInstrument,
-          quantity: Decimal(string: "750.00")!, type: .transfer),
-      ]
-    )
-
-    let created = try await repository.create(original)
-
-    var updated = created
-    updated.payee = "Updated Payee"
-
-    let result = try await repository.update(updated)
-
-    // Verify all fields match the updated transaction
-    #expect(result.id == created.id)
-    #expect(result.legs.allSatisfy { $0.type == .transfer })
-    #expect(result.date == date)
-    #expect(result.legs.count == 2)
-    #expect(result.legs[0].accountId == fromAccountId)
-    #expect(result.legs[1].accountId == toAccountId)
-    #expect(result.legs[0].quantity == Decimal(string: "-750.00")!)
-    #expect(result.payee == "Updated Payee")
-    #expect(result.notes == "Some notes")
-    #expect(result.legs.contains(where: { $0.categoryId == categoryId }))
-    #expect(result.legs.contains(where: { $0.earmarkId == earmarkId }))
-    #expect(result.recurPeriod == .month)
-    #expect(result.recurEvery == 2)
-
-    // Verify persistence by fetching back from repository (scheduled filter needed since
-    // the transaction has recurPeriod set, and the default filter excludes scheduled)
-    let page = try await repository.fetch(
-      filter: TransactionFilter(scheduled: true),
-      page: 0,
-      pageSize: 50
-    )
-    let fetched = try #require(page.transactions.first(where: { $0.id == created.id }))
-
-    #expect(fetched.legs.allSatisfy { $0.type == .transfer })
-    #expect(fetched.date == date)
-    #expect(fetched.legs.count == 2)
-    #expect(fetched.legs[0].accountId == fromAccountId)
-    #expect(fetched.legs[1].accountId == toAccountId)
-    #expect(fetched.legs[0].quantity == Decimal(string: "-750.00")!)
-    #expect(fetched.payee == "Updated Payee")
-    #expect(fetched.notes == "Some notes")
-    #expect(fetched.legs.contains(where: { $0.categoryId == categoryId }))
-    #expect(fetched.legs.contains(where: { $0.earmarkId == earmarkId }))
-    #expect(fetched.recurPeriod == .month)
-    #expect(fetched.recurEvery == 2)
-  }
-
-  @Test("clearing filter reloads all")
-  func testClearingFilterReloadsAll() async throws {
-    let repository = makeCloudKitTransactionRepository(initialTransactions: makeTestTransactions())
-    // First apply a filter
-    let filteredPage = try await repository.fetch(
-      filter: TransactionFilter(payee: "Woolworths"),
-      page: 0,
-      pageSize: 50
-    )
-
-    // Then clear the filter
-    let allPage = try await repository.fetch(
-      filter: TransactionFilter(),
-      page: 0,
-      pageSize: 50
-    )
-
-    #expect(allPage.transactions.count >= filteredPage.transactions.count)
-  }
-
-  @Test("priorBalance is sum of transactions before the page")
-  func testPriorBalanceAcrossPages() async throws {
-    let repository = makeCloudKitTransactionRepository(
-      initialTransactions: makePaginationTestTransactions())
-    let accountFilter = TransactionFilter(accountId: paginationTestAccountId)
-    let page0 = try await repository.fetch(
-      filter: accountFilter,
-      page: 0,
-      pageSize: 2
-    )
-
-    let page1 = try await repository.fetch(
-      filter: accountFilter,
-      page: 1,
-      pageSize: 2
-    )
-
-    // priorBalance for page 0 should be sum of transactions on page 1+
-    let page1Sum = page1.transactions.reduce(
-      InstrumentAmount.zero(instrument: .defaultTestInstrument)
-    ) {
-      $0
-        + $1.legs.reduce(InstrumentAmount.zero(instrument: .defaultTestInstrument)) {
-          $0 + $1.amount
-        }
-    }
-    let page1Prior = try #require(page1.priorBalance)
-    let page1PriorSum = page1Sum + page1Prior
-    let page0Prior = try #require(page0.priorBalance)
-
-    #expect(
-      page0Prior == page1PriorSum,
-      "priorBalance of page 0 should equal sum of all older transactions")
-  }
-
-  @Test("empty page returns zero priorBalance")
-  func testEmptyPagePriorBalance() async throws {
-    let repository = makeCloudKitTransactionRepository(
-      initialTransactions: makePaginationTestTransactions())
-    let page = try await repository.fetch(
-      filter: TransactionFilter(),
-      page: 100,
-      pageSize: 10
-    )
-
-    #expect(page.transactions.isEmpty)
-    let prior = try #require(page.priorBalance)
-    #expect(prior.isZero)
-  }
-
-  @Test("priorBalance is labelled with the account's own instrument")
-  func testPriorBalanceUsesAccountInstrument() async throws {
-    // Non-profile instrument for the viewing account.
-    let accountInstrument = Instrument.USD
-    let accountId = UUID()
-    let (backend, container) = try TestBackend.create()
-    let account = Account(
-      id: accountId, name: "USD Account", type: .bank, instrument: accountInstrument)
-    TestBackend.seed(
-      accounts: [(account, InstrumentAmount.zero(instrument: accountInstrument))],
-      in: container)
-    // Two transactions in the USD account so we need a priorBalance across pages.
-    let tx1 = Transaction(
-      date: Date(timeIntervalSince1970: 1),
-      payee: "Older",
-      legs: [
-        TransactionLeg(
-          accountId: accountId, instrument: accountInstrument,
-          quantity: Decimal(string: "10")!, type: .income)
-      ])
-    let tx2 = Transaction(
-      date: Date(timeIntervalSince1970: 2),
-      payee: "Newer",
-      legs: [
-        TransactionLeg(
-          accountId: accountId, instrument: accountInstrument,
-          quantity: Decimal(string: "20")!, type: .income)
-      ])
-    TestBackend.seed(transactions: [tx1, tx2], in: container)
-
-    // Page size of 1 forces a non-zero priorBalance on page 0.
-    let page0 = try await backend.transactions.fetch(
-      filter: TransactionFilter(accountId: accountId), page: 0, pageSize: 1)
-    #expect(page0.targetInstrument == accountInstrument)
-
-    // Empty paged-past-end response should also use the account's instrument.
-    let pageN = try await backend.transactions.fetch(
-      filter: TransactionFilter(accountId: accountId), page: 99, pageSize: 1)
-    #expect(pageN.targetInstrument == accountInstrument)
-  }
-
-  @Test("priorBalance converts multi-instrument legs to account instrument at today's rate")
-  func testPriorBalanceMultiInstrumentConverts() async throws {
-    // Account is AUD. Historic transactions left three legs behind (one still
-    // in a different instrument) — a trade-style split.
-    let accountInstrument = Instrument.AUD
-    let foreignInstrument = Instrument.USD
-    let accountId = UUID()
-
-    // USD -> AUD at 1.5 today. FixedRateClient keys are ISO-8601 date →
-    // quote-currency rates (base is passed separately to fetchRates).
-    let todayFormatter = ISO8601DateFormatter()
-    todayFormatter.formatOptions = [.withFullDate]
-    let todayKey = todayFormatter.string(from: Date())
-    let rates: [String: [String: Decimal]] = [
-      todayKey: ["AUD": Decimal(string: "1.5")!]
-    ]
-    let (backend, container) = try TestBackend.create(
-      instrument: accountInstrument, exchangeRates: rates)
-    let account = Account(
-      id: accountId, name: "Brokerage", type: .bank, instrument: accountInstrument)
-    TestBackend.seed(
-      accounts: [(account, InstrumentAmount.zero(instrument: accountInstrument))],
-      in: container)
-
-    // Two older transactions (will be on page 1+, contributing to priorBalance).
-    // tx1: AUD +100 (cash deposit)
-    let tx1 = Transaction(
-      date: Date(timeIntervalSince1970: 1_000_000),
-      payee: "Cash in",
-      legs: [
-        TransactionLeg(
-          accountId: accountId, instrument: accountInstrument,
-          quantity: Decimal(100), type: .income)
-      ])
-    // tx2: USD +20 (foreign cash in — will be converted @ 1.5 = +30 AUD today)
-    let tx2 = Transaction(
-      date: Date(timeIntervalSince1970: 1_000_001),
-      payee: "Foreign in",
-      legs: [
-        TransactionLeg(
-          accountId: accountId, instrument: foreignInstrument,
-          quantity: Decimal(20), type: .income)
-      ])
-    // tx3 (newest): a single page-0 entry so priorBalance covers tx1+tx2.
-    let tx3 = Transaction(
-      date: Date(timeIntervalSince1970: 2_000_000),
-      payee: "Today",
-      legs: [
-        TransactionLeg(
-          accountId: accountId, instrument: accountInstrument,
-          quantity: Decimal(5), type: .income)
-      ])
-    TestBackend.seed(transactions: [tx1, tx2, tx3], in: container)
-
-    // pageSize: 1 => page 0 has only tx3; priorBalance = tx1 + tx2(USD->AUD).
-    let page0 = try await backend.transactions.fetch(
-      filter: TransactionFilter(accountId: accountId), page: 0, pageSize: 1)
-
-    let prior = try #require(page0.priorBalance)
-    #expect(prior.instrument == accountInstrument)
-    // 100 AUD + 20 USD @ 1.5 = 130 AUD.
-    #expect(prior == InstrumentAmount(quantity: Decimal(130), instrument: accountInstrument))
-    #expect(page0.targetInstrument == accountInstrument)
-  }
-
-  @Test("priorBalance is nil when conversion fails for any foreign leg")
-  func testPriorBalanceNilOnConversionFailure() async throws {
-    // Account AUD; historic leg is in an unsupported pair (no rate provided).
-    let accountInstrument = Instrument.AUD
-    let foreignInstrument = Instrument.USD
-    let accountId = UUID()
-    // Empty rate table => USD->AUD lookup fails.
-    let (backend, container) = try TestBackend.create(
-      instrument: accountInstrument, exchangeRates: [:])
-    let account = Account(
-      id: accountId, name: "Brokerage", type: .bank, instrument: accountInstrument)
-    TestBackend.seed(
-      accounts: [(account, InstrumentAmount.zero(instrument: accountInstrument))],
-      in: container)
-
-    let tx1 = Transaction(
-      date: Date(timeIntervalSince1970: 1_000_000),
-      payee: "Foreign in",
-      legs: [
-        TransactionLeg(
-          accountId: accountId, instrument: foreignInstrument,
-          quantity: Decimal(20), type: .income)
-      ])
-    let tx2 = Transaction(
-      date: Date(timeIntervalSince1970: 2_000_000),
-      payee: "Today",
-      legs: [
-        TransactionLeg(
-          accountId: accountId, instrument: accountInstrument,
-          quantity: Decimal(5), type: .income)
-      ])
-    TestBackend.seed(transactions: [tx1, tx2], in: container)
-
-    let page0 = try await backend.transactions.fetch(
-      filter: TransactionFilter(accountId: accountId), page: 0, pageSize: 1)
-
-    #expect(page0.priorBalance == nil, "conversion failure should null the prior balance")
-    #expect(page0.targetInstrument == accountInstrument)
-    // Transactions still flow through — failure degrades gracefully.
-    #expect(page0.transactions.count == 1)
-  }
-
-  @Test("priorBalance skips conversion when all legs share the account instrument")
-  func testPriorBalanceSingleInstrumentNoConversionNeeded() async throws {
-    // Empty rate table: if conversion were invoked it would throw. Test asserts
-    // the same-instrument short-circuit keeps working.
-    let accountInstrument = Instrument.AUD
-    let accountId = UUID()
-    let (backend, container) = try TestBackend.create(
-      instrument: accountInstrument, exchangeRates: [:])
-    let account = Account(
-      id: accountId, name: "Bank", type: .bank, instrument: accountInstrument)
-    TestBackend.seed(
-      accounts: [(account, InstrumentAmount.zero(instrument: accountInstrument))],
-      in: container)
-
-    let tx1 = Transaction(
-      date: Date(timeIntervalSince1970: 1_000_000),
-      payee: "Older",
-      legs: [
-        TransactionLeg(
-          accountId: accountId, instrument: accountInstrument,
-          quantity: Decimal(50), type: .income)
-      ])
-    let tx2 = Transaction(
-      date: Date(timeIntervalSince1970: 2_000_000),
-      payee: "Newer",
-      legs: [
-        TransactionLeg(
-          accountId: accountId, instrument: accountInstrument,
-          quantity: Decimal(25), type: .income)
-      ])
-    TestBackend.seed(transactions: [tx1, tx2], in: container)
-
-    let page0 = try await backend.transactions.fetch(
-      filter: TransactionFilter(accountId: accountId), page: 0, pageSize: 1)
-
-    let prior = try #require(page0.priorBalance)
-    #expect(prior == InstrumentAmount(quantity: Decimal(50), instrument: accountInstrument))
-  }
-
-  @Test("transfer creates with two legs")
-  func testTransferCreatesTwoLegs() async throws {
-    let repository = makeCloudKitTransactionRepository()
-    let fromAccount = UUID()
-    let toAccount = UUID()
-    let transfer = Transaction(
-      date: Date(),
-      payee: "Transfer",
-      legs: [
-        TransactionLeg(
-          accountId: fromAccount, instrument: .defaultTestInstrument,
-          quantity: Decimal(string: "-5.00")!, type: .transfer),
-        TransactionLeg(
-          accountId: toAccount, instrument: .defaultTestInstrument,
-          quantity: Decimal(string: "5.00")!, type: .transfer),
-      ]
-    )
-
-    let created = try await repository.create(transfer)
-    #expect(created.legs.count == 2)
-    #expect(created.isTransfer)
-  }
-
-  // MARK: - Multi-instrument persistence
-
-  @Test("currency conversion transfer persists legs in distinct instruments")
-  func testCurrencyConversionPersistsLegsInDistinctInstruments() async throws {
-    let repository = makeCloudKitTransactionRepository()
-    let accountId = UUID()
-    let date = Calendar.current.date(from: DateComponents(year: 2026, month: 3, day: 15))!
-    let conversion = Transaction(
-      date: date,
-      payee: "FX",
-      legs: [
-        TransactionLeg(
-          accountId: accountId, instrument: .AUD,
-          quantity: Decimal(string: "-1000.00")!, type: .transfer),
-        TransactionLeg(
-          accountId: accountId, instrument: .USD,
-          quantity: Decimal(string: "650.00")!, type: .transfer),
-      ]
-    )
-
-    let created = try await repository.create(conversion)
-    #expect(created.legs.count == 2)
-
-    let page = try await repository.fetch(
-      filter: TransactionFilter(accountId: accountId),
-      page: 0,
-      pageSize: 10
-    )
-    let fetched = try #require(page.transactions.first(where: { $0.id == created.id }))
-    #expect(fetched.legs.count == 2)
-    let audLeg = try #require(fetched.legs.first(where: { $0.instrument == .AUD }))
-    let usdLeg = try #require(fetched.legs.first(where: { $0.instrument == .USD }))
-    #expect(audLeg.quantity == Decimal(string: "-1000.00")!)
-    #expect(usdLeg.quantity == Decimal(string: "650.00")!)
-  }
-
-  @Test("stock trade transaction persists fiat and stock legs")
-  func testStockTradePersistsFiatAndStockLegs() async throws {
-    let repository = makeCloudKitTransactionRepository()
-    let accountId = UUID()
-    let bhp = Instrument.stock(ticker: "BHP.AX", exchange: "ASX", name: "BHP")
-    let trade = Transaction(
-      date: Date(),
-      payee: "Buy 150 BHP",
-      legs: [
-        TransactionLeg(
-          accountId: accountId, instrument: .AUD,
-          quantity: Decimal(string: "-6345.00")!, type: .transfer),
-        TransactionLeg(
-          accountId: accountId, instrument: bhp, quantity: Decimal(150), type: .transfer),
-      ]
-    )
-
-    let created = try await repository.create(trade)
-    #expect(created.legs.count == 2)
-
-    let page = try await repository.fetch(
-      filter: TransactionFilter(accountId: accountId),
-      page: 0,
-      pageSize: 10
-    )
-    let fetched = try #require(page.transactions.first(where: { $0.id == created.id }))
-    let audLeg = try #require(fetched.legs.first(where: { $0.instrument == .AUD }))
-    let stockLeg = try #require(fetched.legs.first(where: { $0.instrument.kind == .stock }))
-    #expect(audLeg.quantity == Decimal(string: "-6345.00")!)
-    #expect(stockLeg.instrument == bhp)
-    #expect(stockLeg.quantity == Decimal(150))
-  }
-
-  @Test("three-leg trade with fee in third instrument persists each leg")
-  func testThreeLegTradeWithForeignFeePersistsAllLegs() async throws {
-    let repository = makeCloudKitTransactionRepository()
-    let accountId = UUID()
-    let feeCategoryId = UUID()
-    let aapl = Instrument.stock(ticker: "AAPL", exchange: "NASDAQ", name: "Apple")
-    // Sell USD, buy AAPL, fee in AUD — three distinct instruments across legs.
-    let trade = Transaction(
-      date: Date(),
-      payee: "Buy 10 Apple",
-      legs: [
-        TransactionLeg(
-          accountId: accountId, instrument: .USD,
-          quantity: Decimal(string: "-1855.00")!, type: .transfer),
-        TransactionLeg(
-          accountId: accountId, instrument: aapl, quantity: Decimal(10), type: .transfer),
-        TransactionLeg(
-          accountId: accountId, instrument: .AUD,
-          quantity: Decimal(string: "-7.50")!, type: .expense,
-          categoryId: feeCategoryId),
-      ]
-    )
-
-    let created = try await repository.create(trade)
-    #expect(created.legs.count == 3)
-
-    let page = try await repository.fetch(
-      filter: TransactionFilter(accountId: accountId),
-      page: 0,
-      pageSize: 10
-    )
-    let fetched = try #require(page.transactions.first(where: { $0.id == created.id }))
-    #expect(fetched.legs.count == 3)
-    let instrumentIds = Set(fetched.legs.map { $0.instrument.id })
-    #expect(instrumentIds == [Instrument.USD.id, aapl.id, Instrument.AUD.id])
-    let feeLeg = try #require(fetched.legs.first(where: { $0.type == .expense }))
-    #expect(feeLeg.instrument == .AUD)
-    #expect(feeLeg.categoryId == feeCategoryId)
-  }
-
-  @Test("filter by accountId returns transactions regardless of leg instrument")
-  func testFilterByAccountReturnsMultiInstrumentTransactions() async throws {
-    let accountId = UUID()
-    let bhp = Instrument.stock(ticker: "BHP.AX", exchange: "ASX", name: "BHP")
-    let initial: [Transaction] = [
-      Transaction(
-        date: Calendar.current.date(from: DateComponents(year: 2026, month: 1, day: 5))!,
-        payee: "Opening AUD",
-        legs: [
-          TransactionLeg(
-            accountId: accountId, instrument: .AUD,
-            quantity: Decimal(string: "10000.00")!, type: .openingBalance)
-        ]),
-      Transaction(
-        date: Calendar.current.date(from: DateComponents(year: 2026, month: 1, day: 10))!,
-        payee: "FX",
-        legs: [
-          TransactionLeg(
-            accountId: accountId, instrument: .AUD,
-            quantity: Decimal(string: "-1000.00")!, type: .transfer),
-          TransactionLeg(
-            accountId: accountId, instrument: .USD,
-            quantity: Decimal(string: "650.00")!, type: .transfer),
-        ]),
-      Transaction(
-        date: Calendar.current.date(from: DateComponents(year: 2026, month: 1, day: 15))!,
-        payee: "Buy 100 BHP",
-        legs: [
-          TransactionLeg(
-            accountId: accountId, instrument: .AUD,
-            quantity: Decimal(string: "-4230.00")!, type: .transfer),
-          TransactionLeg(
-            accountId: accountId, instrument: bhp, quantity: Decimal(100), type: .transfer),
-        ]),
-    ]
-    let repository = makeCloudKitTransactionRepository(initialTransactions: initial)
-
-    let page = try await repository.fetch(
-      filter: TransactionFilter(accountId: accountId),
-      page: 0,
-      pageSize: 50
-    )
-    #expect(page.transactions.count == 3)
-    // Union of instruments across all legs on this account must include fiat + stock.
-    let allInstruments = Set(page.transactions.flatMap { $0.legs.map(\.instrument.id) })
-    #expect(allInstruments.contains(Instrument.AUD.id))
-    #expect(allInstruments.contains(Instrument.USD.id))
-    #expect(allInstruments.contains(bhp.id))
-  }
-
-  @Test("transactions are sorted by date descending")
-  func testTransactionsSortedByDateDesc() async throws {
-    let repository = makeCloudKitTransactionRepository(initialTransactions: makeTestTransactions())
-    let page = try await repository.fetch(
-      filter: TransactionFilter(),
-      page: 0,
-      pageSize: 50
-    )
-
-    for i in 0..<(page.transactions.count - 1) {
-      #expect(
-        page.transactions[i].date >= page.transactions[i + 1].date,
-        "Transactions should be sorted by date descending"
-      )
-    }
-  }
-
-  // MARK: - Scheduled Filter
-
-  @Test("filters scheduled transactions")
-  func testFiltersByScheduled() async throws {
-    let repository = makeCloudKitTransactionRepository(
-      initialTransactions: makeScheduledTestTransactions())
-    let scheduledPage = try await repository.fetch(
-      filter: TransactionFilter(scheduled: true),
-      page: 0,
-      pageSize: 50
-    )
-    #expect(scheduledPage.transactions.count == 1)
-    #expect(scheduledPage.transactions[0].isScheduled)
-
-    let nonScheduledPage = try await repository.fetch(
-      filter: TransactionFilter(scheduled: false),
-      page: 0,
-      pageSize: 50
-    )
-    #expect(nonScheduledPage.transactions.count == 2)
-    for txn in nonScheduledPage.transactions {
-      #expect(!txn.isScheduled)
-    }
-  }
-
-  // MARK: - Earmark Filter
-
-  @Test("filters by earmarkId")
-  func testFiltersByEarmarkId() async throws {
-    let repository = makeCloudKitTransactionRepository(initialTransactions: makeTestTransactions())
-    // The test data has one earmarked transaction
-    let allPage = try await repository.fetch(
-      filter: TransactionFilter(),
-      page: 0,
-      pageSize: 50
-    )
-    let earmarked = allPage.transactions.filter { $0.legs.contains(where: { $0.earmarkId != nil }) }
-    #expect(earmarked.count == 1, "Test data should have one earmarked transaction")
-
-    let earmarkId = earmarked[0].legs.first(where: { $0.earmarkId != nil })!.earmarkId!
-    let filteredPage = try await repository.fetch(
-      filter: TransactionFilter(earmarkId: earmarkId),
-      page: 0,
-      pageSize: 50
-    )
-
-    #expect(filteredPage.transactions.count == 1)
-    #expect(filteredPage.transactions[0].legs.contains(where: { $0.earmarkId == earmarkId }))
-  }
-
-  // MARK: - Account Filter
-
-  @Test("filters by accountId including transfers")
-  func testFiltersByAccountIdIncludingTransfers() async throws {
-    let repository = makeCloudKitTransactionRepository(
-      initialTransactions: makeTransferTestTransactions())
-    let sourceAccountId = transferSourceAccountId
-    let destAccountId = transferDestAccountId
-
-    // Filter by source account — should include the transfer AND the expense
-    let sourcePage = try await repository.fetch(
-      filter: TransactionFilter(accountId: sourceAccountId),
-      page: 0,
-      pageSize: 50
-    )
-    #expect(sourcePage.transactions.count == 2)
-
-    // Filter by dest account — should include the transfer only
-    let destPage = try await repository.fetch(
-      filter: TransactionFilter(accountId: destAccountId),
-      page: 0,
-      pageSize: 50
-    )
-    #expect(destPage.transactions.count == 1)
-    #expect(destPage.transactions[0].legs.allSatisfy { $0.type == .transfer })
-  }
-
-  // MARK: - Payee Suggestions
-
-  @Test("fetchPayeeSuggestions returns prefix matches sorted by frequency")
-  func testPayeeSuggestions() async throws {
-    let repository = makeCloudKitTransactionRepository(
-      initialTransactions: makePayeeSuggestionTestTransactions())
-    let suggestions = try await repository.fetchPayeeSuggestions(prefix: "Wool")
-
-    #expect(suggestions.count == 1)
-    #expect(suggestions[0] == "Woolworths")
-  }
-
-  @Test("fetchPayeeSuggestions is case insensitive")
-  func testPayeeSuggestionsCaseInsensitive() async throws {
-    let repository = makeCloudKitTransactionRepository(
-      initialTransactions: makePayeeSuggestionTestTransactions())
-    let suggestions = try await repository.fetchPayeeSuggestions(prefix: "wool")
-
-    #expect(suggestions.count == 1)
-    #expect(suggestions[0] == "Woolworths")
-  }
-
-  @Test("fetchPayeeSuggestions returns empty for empty prefix")
-  func testPayeeSuggestionsEmptyPrefix() async throws {
-    let repository = makeCloudKitTransactionRepository(
-      initialTransactions: makePayeeSuggestionTestTransactions())
-    let suggestions = try await repository.fetchPayeeSuggestions(prefix: "")
-    #expect(suggestions.isEmpty)
-  }
-
-  @Test("fetchPayeeSuggestions sorts by frequency")
-  func testPayeeSuggestionsByFrequency() async throws {
-    let repository = makeCloudKitTransactionRepository(
-      initialTransactions: makePayeeSuggestionTestTransactions())
-    // "Coles" appears 3 times, "Coffee Shop" appears 1 time
-    let suggestions = try await repository.fetchPayeeSuggestions(prefix: "Co")
-
-    #expect(suggestions.count == 2)
-    #expect(suggestions[0] == "Coles", "Most frequent payee should be first")
-    #expect(suggestions[1] == "Coffee Shop")
-  }
-}
-
-// MARK: - Test Data Helpers
-
-private func makeLeg(
-  accountId: UUID,
-  quantity: Decimal,
-  type: TransactionType,
-  categoryId: UUID? = nil,
-  earmarkId: UUID? = nil
-) -> TransactionLeg {
-  TransactionLeg(
-    accountId: accountId,
-    instrument: .defaultTestInstrument,
-    quantity: quantity,
-    type: type,
-    categoryId: categoryId,
-    earmarkId: earmarkId
-  )
-}
-
-// Helper function to create test transactions with various attributes
-private func makeTestTransactions() -> [Transaction] {
-  let accountId = UUID()
-  let groceryCategoryId = UUID()
-  let transportCategoryId = UUID()
-  let earmarkId = UUID()
-  let calendar = Calendar.current
-
-  let transactions: [Transaction] = [
-    // Grocery expense in June
-    Transaction(
-      date: calendar.date(from: DateComponents(year: 2024, month: 6, day: 15))!,
-      payee: "Woolworths",
-      legs: [
-        makeLeg(
-          accountId: accountId, quantity: Decimal(string: "-50.23")!, type: .expense,
-          categoryId: groceryCategoryId)
-      ]
-    ),
-    // Transport expense in July
-    Transaction(
-      date: calendar.date(from: DateComponents(year: 2024, month: 7, day: 10))!,
-      payee: "Metro Transport",
-      legs: [
-        makeLeg(
-          accountId: accountId, quantity: Decimal(string: "-35.00")!, type: .expense,
-          categoryId: transportCategoryId)
-      ]
-    ),
-    // Income in May
-    Transaction(
-      date: calendar.date(from: DateComponents(year: 2024, month: 5, day: 30))!,
-      payee: "Employer Pty Ltd",
-      legs: [makeLeg(accountId: accountId, quantity: Decimal(string: "3500.00")!, type: .income)]
-    ),
-    // Grocery expense in April (older, different payee)
-    Transaction(
-      date: calendar.date(from: DateComponents(year: 2024, month: 4, day: 20))!,
-      payee: "Coles",
-      legs: [
-        makeLeg(
-          accountId: accountId, quantity: Decimal(string: "-42.00")!, type: .expense,
-          categoryId: groceryCategoryId)
-      ]
-    ),
-    // Earmarked expense in June
-    Transaction(
-      date: calendar.date(from: DateComponents(year: 2024, month: 6, day: 20))!,
-      payee: "Electronics Store",
-      legs: [
-        makeLeg(
-          accountId: accountId, quantity: Decimal(string: "-100.00")!, type: .expense,
-          categoryId: transportCategoryId, earmarkId: earmarkId)
-      ]
-    ),
-  ]
-
-  return transactions
-}
-
-private let paginationTestAccountId = UUID()
-
-private func makePaginationTestTransactions() -> [Transaction] {
-  let accountId = paginationTestAccountId
-  let calendar = Calendar.current
-  return [
-    Transaction(
-      date: calendar.date(from: DateComponents(year: 2024, month: 1, day: 1))!,
-      payee: "Jan Income",
-      legs: [makeLeg(accountId: accountId, quantity: Decimal(string: "10.00")!, type: .income)]
-    ),
-    Transaction(
-      date: calendar.date(from: DateComponents(year: 2024, month: 2, day: 1))!,
-      payee: "Feb Expense",
-      legs: [makeLeg(accountId: accountId, quantity: Decimal(string: "-3.00")!, type: .expense)]
-    ),
-    Transaction(
-      date: calendar.date(from: DateComponents(year: 2024, month: 3, day: 1))!,
-      payee: "Mar Income",
-      legs: [makeLeg(accountId: accountId, quantity: Decimal(string: "20.00")!, type: .income)]
-    ),
-    Transaction(
-      date: calendar.date(from: DateComponents(year: 2024, month: 4, day: 1))!,
-      payee: "Apr Expense",
-      legs: [makeLeg(accountId: accountId, quantity: Decimal(string: "-5.00")!, type: .expense)]
-    ),
-  ]
-}
-
-private func makeScheduledTestTransactions() -> [Transaction] {
-  let accountId = UUID()
-  let calendar = Calendar.current
-  return [
-    // Non-scheduled expense
-    Transaction(
-      date: calendar.date(from: DateComponents(year: 2024, month: 6, day: 15))!,
-      payee: "Store",
-      legs: [makeLeg(accountId: accountId, quantity: Decimal(string: "-50.00")!, type: .expense)]
-    ),
-    // Non-scheduled income
-    Transaction(
-      date: calendar.date(from: DateComponents(year: 2024, month: 6, day: 1))!,
-      payee: "Salary",
-      legs: [makeLeg(accountId: accountId, quantity: Decimal(string: "1000.00")!, type: .income)]
-    ),
-    // Scheduled (recurring) transaction
-    Transaction(
-      date: calendar.date(from: DateComponents(year: 2024, month: 7, day: 1))!,
-      payee: "Netflix",
-      recurPeriod: .month,
-      recurEvery: 1,
-      legs: [makeLeg(accountId: accountId, quantity: Decimal(string: "-20.00")!, type: .expense)]
-    ),
-  ]
-}
-
-// Stable IDs for transfer test data (shared between seed and assertions)
-private let transferSourceAccountId = UUID()
-private let transferDestAccountId = UUID()
-
-private func makeTransferTestTransactions() -> [Transaction] {
-  let sourceAccountId = transferSourceAccountId
-  let destAccountId = transferDestAccountId
-  let calendar = Calendar.current
-  return [
-    // Transfer from source to dest
-    Transaction(
-      date: calendar.date(from: DateComponents(year: 2024, month: 6, day: 15))!,
-      payee: "Transfer",
-      legs: [
-        makeLeg(accountId: sourceAccountId, quantity: Decimal(string: "-100.00")!, type: .transfer),
-        makeLeg(accountId: destAccountId, quantity: Decimal(string: "100.00")!, type: .transfer),
-      ]
-    ),
-    // Expense on source account only
-    Transaction(
-      date: calendar.date(from: DateComponents(year: 2024, month: 6, day: 10))!,
-      payee: "Coffee",
-      legs: [
-        makeLeg(accountId: sourceAccountId, quantity: Decimal(string: "-50.00")!, type: .expense)
-      ]
-    ),
-  ]
-}
-
-private func makePayeeSuggestionTestTransactions() -> [Transaction] {
-  let accountId = UUID()
-  let calendar = Calendar.current
-  return [
-    Transaction(
-      date: calendar.date(from: DateComponents(year: 2024, month: 6, day: 1))!,
-      payee: "Woolworths",
-      legs: [makeLeg(accountId: accountId, quantity: Decimal(string: "-10.00")!, type: .expense)]
-    ),
-    Transaction(
-      date: calendar.date(from: DateComponents(year: 2024, month: 6, day: 2))!,
-      payee: "Coles",
-      legs: [makeLeg(accountId: accountId, quantity: Decimal(string: "-20.00")!, type: .expense)]
-    ),
-    Transaction(
-      date: calendar.date(from: DateComponents(year: 2024, month: 6, day: 3))!,
-      payee: "Coles",
-      legs: [makeLeg(accountId: accountId, quantity: Decimal(string: "-30.00")!, type: .expense)]
-    ),
-    Transaction(
-      date: calendar.date(from: DateComponents(year: 2024, month: 6, day: 4))!,
-      payee: "Coles",
-      legs: [makeLeg(accountId: accountId, quantity: Decimal(string: "-15.00")!, type: .expense)]
-    ),
-    Transaction(
-      date: calendar.date(from: DateComponents(year: 2024, month: 6, day: 5))!,
-      payee: "Coffee Shop",
-      legs: [makeLeg(accountId: accountId, quantity: Decimal(string: "-5.00")!, type: .expense)]
-    ),
-  ]
-}
-
-private func makeCloudKitTransactionRepository(
-  initialTransactions: [Transaction] = [],
-  instrument: Instrument = .defaultTestInstrument,
-  exchangeRates: [String: [String: Decimal]] = [:]
-) -> CloudKitTransactionRepository {
-  let container = try! TestModelContainer.create()
-  let rateClient = FixedRateClient(rates: exchangeRates)
-  let cacheDir = FileManager.default.temporaryDirectory
-    .appendingPathComponent("test-rates-\(UUID().uuidString)")
-  let exchangeRateService = ExchangeRateService(client: rateClient, cacheDirectory: cacheDir)
-  let conversionService = FiatConversionService(exchangeRates: exchangeRateService)
-  let repo = CloudKitTransactionRepository(
-    modelContainer: container,
-    instrument: instrument,
-    conversionService: conversionService)
-
-  if !initialTransactions.isEmpty {
-    let context = ModelContext(container)
-    for txn in initialTransactions {
-      context.insert(TransactionRecord.from(txn))
-      for (index, leg) in txn.legs.enumerated() {
-        context.insert(TransactionLegRecord.from(leg, transactionId: txn.id, sortOrder: index))
-      }
-    }
-    try! context.save()
-  }
-
-  return repo
-}
